:truck: Move into pages module
