+++ src/app/collection/pages/cyberskills-dashboard/components/csv-gen-modal/csv-gen-modal.component.scss
@@ -1,5 +1,5 @@
 @import "vars.scss";
-@import "../../../collection-globals.scss";
+@import "../../../../collection-globals.scss";
 
 .modal-container {
     margin: 0 auto;
