+++ src/app/collection/pages/cyberskills-dashboard/components/cyberskills-card/cyberskills-card.component.scss
