+++ src/app/collection/pages/cyberskills-dashboard/components/cyberskills-filters/cyberskills-filters.component.scss
